% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemographicTable.R
\name{as_flextable.DemographicTable}
\alias{as_flextable.DemographicTable}
\title{Convert \link{DemographicTable} to \link[flextable]{flextable}}
\usage{
\method{as_flextable}{DemographicTable}(x, ...)
}
\arguments{
\item{x}{a \link{DemographicTable} object}

\item{...}{potential additional parameters, not currently in use}
}
\value{
Function \link{as_flextable.DemographicTable} returns a \link[flextable]{flextable} object.
}
\description{
Convert a \link{DemographicTable} to \link[flextable]{flextable} object.
}
\note{
End user may use \link[flextable]{set_caption} to add a caption to the output demographic table.
}
