% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{c.DemographicTable}
\alias{c.DemographicTable}
\title{Concatenate \link{DemographicTable} Objects}
\usage{
\method{c}{DemographicTable}(...)
}
\arguments{
\item{...}{one or more \link{DemographicTable} objects}
}
\value{
Function \code{\link[=c.DemographicTable]{c.DemographicTable()}} returns a \link{DemographicTable} object.
}
\description{
Concatenate \link{DemographicTable} Objects
}
\keyword{internal}
