% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class1List.R
\name{class1List}
\alias{class1List}
\title{class1List}
\usage{
class1List(data)
}
\arguments{
\item{data}{a \code{\link[base]{data.frame}}}
}
\value{
\code{\link{class1List}} returns a \code{\link[base]{list}} 
of the first \code{\link[base]{class}} of each columns in the input \code{\link[base]{data.frame}}
}
\description{
List the first \code{\link[base]{class}} of each columns in a \code{\link[base]{data.frame}}
}
\examples{
class1List(esoph)

}
