% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustColWidths.R
\name{colWidth}
\alias{colWidth}
\title{Text Width (for \pkg{flextable})}
\usage{
colWidth(x, linefeed = TRUE)
}
\arguments{
\item{x}{An R object convertible to \link[base]{data.frame}}

\item{linefeed}{see \code{\link{nchar_lf}}}
}
\value{
\code{\link{colWidth}} returns an \link[base]{integer} vector
of the print widths of a \link[base]{data.frame}
\strong{not} considering the row names.
}
\description{
Determine the text width of screen output
}
\examples{
head(mtcars)
colWidth(mtcars)

}
