% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{DemographicTable-package}
\alias{DemographicTable-package}
\title{DemographicTable: Create Demographic Table}
\description{
To create demographic table with simple summary statistics, with optional comparison(s) over one or more groups.
}
\note{
Help files of individual functions are intentionally suppressed in the \code{pdf} manual.
Users are encouraged to get started with

\code{vignette('intro', package = 'DemographicTable')}
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com}

Authors:
\itemize{
  \item Tingting Zhan \email{tingtingzhan@gmail.com}
}

}
