% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemographicTable.R
\name{DemographicTable}
\alias{DemographicTable}
\title{Demographic Table}
\usage{
DemographicTable(data, data.name = substitute(data), by = NULL, ...)
}
\arguments{
\item{data}{a \link[base]{data.frame}}

\item{data.name}{\link[base]{character} scalar,
or the argument \link[base]{call} of \code{data}.
A user-friendly name of the input \code{data}.}

\item{by}{one-sided \link[stats]{formula},
the name(s) of sub-group(s) for which the summary statistics are provided.
Default \code{NULL} indicating no sub-groups.}

\item{...}{additional parameters, currently not in use}
}
\value{
The function \code{\link[=DemographicTable]{DemographicTable()}} returns an object of S3 class \code{'DemographicTable'},
which is a \link[stats]{listof} \code{'sumtab'} elements.
}
\description{
To create a demographic table with simple summary statistics,
with optional comparison(s) over one or more groups.
}
\details{
A demographic table with simple summary statistics, with optional comparison(s) over one or more groups, is created.

\link[base]{numeric} variables are summarized in means, standard deviations, medians, inter-quartile-ranges (IQR),
\eqn{p}-value of Shapiro-Wilk normality test and ranges.
If \code{group} is specified, they are compared using two-sample \link[stats]{t.test},
\link[stats]{wilcox.test} (Wilcoxon / Mann-Whitney), one-way \link[stats]{aov} (ANOVA) and/or
\link[stats]{kruskal.test} (Kruskal-Wallis).

\link[base]{logical} and \link[base]{factor} variables are summarized in counts and percentages.
If \code{group} is specified, they are compared using \link[stats]{prop.test} (chi-squared)
and/or \link[stats]{fisher.test} (Fisher's exact).
}
\keyword{internal}
