% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg_}
\alias{msg_}
\alias{msg_logical}
\title{Various \link[base]{warning} and/or \link[base]{stop} Messages}
\usage{
msg_logical()
}
\description{
Various \link[base]{warning} and/or \link[base]{stop} Messages
}
\examples{
cat(msg_logical())
}
\keyword{internal}
