% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg_logical}
\alias{msg_logical}
\title{\link[base]{warning} Messages on the use of \link[base]{logical} variable}
\usage{
msg_logical()
}
\description{
\link[base]{warning} Messages on the use of \link[base]{logical} variable
}
\examples{
msg_logical()

}
\keyword{internal}
