% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flextable.R
\name{print.DemographicTable}
\alias{print.DemographicTable}
\title{Print \link{DemographicTable}}
\usage{
\method{print}{DemographicTable}(x, ...)
}
\arguments{
\item{x}{a \link{DemographicTable}}

\item{...}{additional parameters of function \link[flextable]{print.flextable}}
}
\value{
Function \code{\link[=print.DemographicTable]{print.DemographicTable()}} returns a \link[flextable]{flextable} \link[base]{invisible}-y.
}
\description{
Print \link{DemographicTable}
}
\keyword{internal}
