% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_.DemographicTable.R
\name{rmd_.DemographicTable}
\alias{rmd_.DemographicTable}
\title{Create R Markdown Script for \link{DemographicTable}}
\usage{
rmd_.DemographicTable(x, xnm, type, ...)
}
\arguments{
\item{x}{a \link{DemographicTable}}

\item{xnm}{\link[base]{language} or \link[base]{character} scalar, call of \code{x}}

\item{type}{..}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{rmd_.DemographicTable} returns a \link[base]{character} \link[base]{vector}.
}
\description{
Method dispatch to \link{DemographicTable} for S3 generic \code{rmd_} (in a different master package).
}
