% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sprintf.DemographicTable.R
\name{Sprintf.DemographicTable}
\alias{Sprintf.DemographicTable}
\title{Short Paragraph to Describe a \link{DemographicTable}}
\usage{
Sprintf.DemographicTable(model, ...)
}
\arguments{
\item{model}{a \link{DemographicTable}}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{Sprintf.DemographicTable} returns a \link[base]{noquote} \link[base]{character} scalar.
}
\description{
To create a short paragraph to describe a \link{DemographicTable}
}
\examples{
Sprintf.DemographicTable(DemographicTable(esoph))
Sprintf.DemographicTable(DemographicTable(ToothGrowth, groups = 'supp', include = 'len'))

}
