% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat.R
\name{sumstat}
\alias{sumstat}
\alias{.sumstat}
\alias{.sumstat.default}
\alias{.sumstat.factor}
\alias{.sumstat.character}
\alias{.sumstat.logical}
\title{Summary Statistics}
\usage{
.sumstat(x, fmt, ...)

\method{.sumstat}{default}(x, fmt = "\%.2f", ...)

\method{.sumstat}{factor}(x, ...)

\method{.sumstat}{character}(x, ...)

\method{.sumstat}{logical}(x, ...)
}
\arguments{
\item{x}{an R object}

\item{fmt}{(optional) \link[base]{character} scalar, only for the function \code{.sumstat.default()}, see function \link[base]{sprintf}}

\item{...}{additional parameters, currently not in use}
}
\value{
The function \code{\link[=.sumstat]{.sumstat()}} returns a \link[base]{character} scalar.
}
\description{
Summary statistics to be presented in a \link{DemographicTable}.
}
\details{
The function \code{.sumstat.default()} accepts all R objects of \link[base]{typeof} \link[base]{double} and \link[base]{integer}.
Statistics of \link[base]{mean}, \link[stats]{sd} and \link[base]{range} are reported.
If the normality assumption via \link[stats]{shapiro.test} is not satisfied, then \link[stats]{median} and \link[stats]{mad} are also reported.
}
\examples{
airquality$Solar.R |> .sumstat()
state.region |> .sumstat()

}
\keyword{internal}
