% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemographicTable.R
\name{as_flextable.DemographicTable}
\alias{as_flextable.DemographicTable}
\title{Convert \code{\link{DemographicTable}} to \code{\link[flextable]{flextable}}}
\usage{
\method{as_flextable}{DemographicTable}(x, font.size = 8, caption, ...)
}
\arguments{
\item{x}{a \code{\link{DemographicTable}}}

\item{font.size}{\link[base]{integer} scalar, the font size (default 8).
See \code{\link[flextable]{fontsize}}}

\item{caption}{(optional) \link[base]{character} scalar, the table caption.
See \code{\link[flextable]{set_caption}}}

\item{...}{potential additional parameters, not currently in use}
}
\value{
\code{\link{as_flextable.DemographicTable}} returns a \link[flextable]{flextable} object.
}
\description{
Convert a \code{\link{DemographicTable}} to \code{\link[flextable]{flextable}}.
}
\seealso{
\link[flextable]{as_flextable}
}
