% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class1List.R
\name{class1List}
\alias{class1List}
\title{class1List}
\usage{
class1List(x)
}
\arguments{
\item{x}{a \link[base]{data.frame} or \link[base]{list}}
}
\value{
\link{class1List} returns a \link[base]{list}
of the first \link[base]{class} of each element of the input.
}
\description{
The first \link[base]{class} of each columns in a \link[base:is.recursive]{recursive} object
}
\examples{
class1List(esoph)
class1List(lm(Ozone ~ Wind + Temp, data = airquality))

}
