% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemographicTable_info.R
\docType{package}
\name{DemographicTable-package}
\alias{DemographicTable-package}
\title{Create Demographic Table}
\description{
\tabular{ll}{
Package: \tab DemographicTable\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2021-12-15\cr
License: \tab GPL (>= 2)\cr
}
}
\details{
This package presents simple summary statistics for numeric, binary and categorical variables
}
\keyword{package}
