% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemographicTable.R
\name{Sprintf.DemographicTable}
\alias{Sprintf.DemographicTable}
\title{Description for \link[DemographicTable]{DemographicTable}}
\usage{
Sprintf.DemographicTable(x)
}
\arguments{
\item{x}{a \link[DemographicTable]{DemographicTable}}
}
\value{
Function \code{\link[=Sprintf.DemographicTable]{Sprintf.DemographicTable()}} returns a \link[base]{character} scalar.
}
\description{
..
}
\keyword{internal}
