% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemographicTable.R
\name{xtable.DemographicTable}
\alias{xtable.DemographicTable}
\title{Write \link{DemographicTable} to LaTeX}
\usage{
\method{xtable}{DemographicTable}(x, ...)
}
\arguments{
\item{x}{a \link{DemographicTable}}

\item{...}{potential parameters of \link[xtable]{xtable}}
}
\value{
\link{xtable.DemographicTable} returns an \link[xtable]{xtable} object.
}
\description{
Write \link{DemographicTable} to LaTeX.
}
\examples{
(tb = DemographicTable(ToothGrowth, groups = 'supp'))
library(xtable)
print(xtable(tb), sanitize.text.function = identity, 
 sanitize.colnames.function = NULL, include.rownames = FALSE)

}
\seealso{
\link[xtable]{xtable}
}
